통합 테스트 오류 수정(Ide, 마이리스트)

+++ src/api/MyListService.js
@@ -1,9 +1,20 @@
+const accessToken = localStorage.getItem("accessToken");
+
 // 마이리스트 전체 조회
 export const fetchMyLists = async userId => {
   try {
+    if (!accessToken) {
+      console.log("access token is not available");
+      throw new Error(`access token 없음`);
+    }
     console.log("fetchMyLists 호출됨", userId);
     const response = await fetch(
-      `https://localhost:8080/directory?userId=${userId}`,
+      `http://localhost:8081/directory?userId=${userId}`,
+      {
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+        },
+      },
     );
 
     if (!response.ok) {
@@ -13,23 +24,23 @@ export const fetchMyLists = async userId => {
   } catch (error) {
     console.error("서버 요청에 실패했습니다: ", error);
 
-    // test: 서버 요청이 실패했을 때 모의 데이터 반환
-    return [{ directoryId: "1", directoryName: "List 1", problemList: [] }];
-
-    // 실제로는 test 부분 삭제하고 아래 코드 주석 해제
-    // throw error;
+    throw error;
   }
 };
 
 // 마이리스트 추가
 export const fetchAddMyList = async (userId, directoryTitle) => {
   try {
+    if (!accessToken) {
+      console.log("access token is not available");
+      throw new Error(`access token 없음`);
+    }
     console.log("fetchAddMyList 호출됨", userId, directoryTitle);
-    const response = await fetch(`https://localhost:8080/directory`, {
+    const response = await fetch(`http://localhost:8081/directory`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
-        // 'Authorization': Bearer accessToken
+        Authorization: `Bearer ${accessToken}`,
       },
       body: JSON.stringify({ userId, directoryTitle }),
     });
@@ -40,25 +51,24 @@ export const fetchAddMyList = async (userId, directoryTitle) => {
   } catch (error) {
     console.error("fetchAddMyList 서버 요청에 실패했습니다: ", error);
 
-    // test: 서버 요청 실패해도 성공 결과 반환
-    return {
-      success: true,
-      message: "[test] 목록 추가 성공",
-    };
-    // 실제로는 test 부분 삭제하고 아래 코드 주석 해제
-    // throw error;
+    throw error;
   }
 };
 
 // 마이리스트 삭제
 export const fetchDeleteMyList = async (userId, directoryId) => {
   try {
+    console.log(accessToken);
+    if (!accessToken) {
+      console.log("access token is not available");
+      throw new Error(`access token 없음`);
+    }
     console.log("fetchDeleteMyList 호출됨", userId, directoryId);
-    const response = await fetch(`http://localhost:8080/directory`, {
+    const response = await fetch(`http://localhost:8081/directory}`, {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
-        // 'Authorization': Bearer accessToken
+        Authorization: `Bearer ${accessToken}`,
       },
       body: JSON.stringify({ userId, directoryId }),
     });
@@ -69,25 +79,23 @@ export const fetchDeleteMyList = async (userId, directoryId) => {
   } catch (error) {
     console.error("fetchDeleteMyList 서버 요청에 실패했습니다: ", error);
 
-    // test: 서버 요청 실패해도 성공 결과 반환
-    return {
-      success: true,
-      message: "[test] 목록 삭제 성공",
-    };
-    // 실제로는 test 부분 삭제하고 아래 코드 주석 해제
-    // throw error;
+    throw error;
   }
 };
 
 // 마이리스트 문제 추가
 export const fetchAddMyListProblem = async (userId, directoryId, problemId) => {
   try {
+    if (!accessToken) {
+      console.log("access token is not available");
+      throw new Error(`access token 없음`);
+    }
     console.log("fetchAddMyListProblem 호출됨", userId, directoryId, problemId);
-    const response = await fetch(`http://localhost:8080/directory/problem`, {
+    const response = await fetch(`http://localhost:8081/directory/problem`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
-        // 'Authorization': Bearer accessToken
+        Authorization: `Bearer ${accessToken}`,
       },
       body: JSON.stringify({ userId, directoryId, problemId }),
     });
@@ -98,13 +106,7 @@ export const fetchAddMyListProblem = async (userId, directoryId, problemId) => {
   } catch (error) {
     console.error("fetchAddMyListProblem 서버 요청에 실패했습니다: ", error);
 
-    // test: 서버 요청 실패해도 성공 결과 반환
-    return {
-      success: true,
-      message: "[test] 목록 삭제 성공",
-    };
-    // 실제로는 test 부분 삭제하고 아래 코드 주석 해제
-    // throw error;
+    throw error;
   }
 };
 
@@ -116,6 +118,10 @@ export const fetchDeleteMyListProblem = async (
   directoryProblemId,
 ) => {
   try {
+    if (!accessToken) {
+      console.log("access token is not available");
+      throw new Error(`access token 없음`);
+    }
     console.log(
       "fetchDeleteMyListProblem 호출됨",
       userId,
@@ -123,11 +129,11 @@ export const fetchDeleteMyListProblem = async (
       problemId,
       directoryProblemId,
     );
-    const response = await fetch(`http://localhost:8080/directory/problem`, {
+    const response = await fetch(`http://localhost:8081/directory/problem`, {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
-        // 'Authorization': Bearer accessToken
+        Authorization: `Bearer ${accessToken}`,
       },
       body: JSON.stringify({
         userId,
@@ -143,12 +149,6 @@ export const fetchDeleteMyListProblem = async (
   } catch (error) {
     console.error("fetchDeleteMyListProblem 서버 요청에 실패했습니다: ", error);
 
-    // test: 서버 요청 실패해도 성공 결과 반환
-    return {
-      success: true,
-      message: "[test] 목록 삭제 성공",
-    };
-    // 실제로는 test 부분 삭제하고 아래 코드 주석 해제
-    // throw error;
+    throw error;
   }
 };
